增加异步上传文档方法
